:art: refactor: API 경로 리팩토링
