+++ src/main/java/zzandori/zzanmoa/comparison/controller/ComparisonController.java
@@ -1,5 +1,6 @@
 package zzandori.zzanmoa.comparison.controller;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,14 +11,15 @@ import zzandori.zzanmoa.comparison.dto.ComparisonRequestDto;
 import zzandori.zzanmoa.comparison.dto.ComparisonResponseDto;
 import zzandori.zzanmoa.comparison.service.ComparisonService;
 
+@Tag(name = "ComparisonController", description = "시장 가격 비교 응답 컨트롤러")
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/compare")
+@RequestMapping("/market")
 public class ComparisonController {
 
     private final ComparisonService comparisonService;
 
-    @PostMapping
+    @PostMapping("/compare")
     public ResponseEntity<ComparisonResponseDto> comparePrice(@RequestBody ComparisonRequestDto comparisonRequestDto) {
         return ResponseEntity.ok(comparisonService.comparePrice(comparisonRequestDto));
     }
