+++ src/main/java/zzandori/zzanmoa/item/service/ItemService.java
@@ -2,14 +2,13 @@ package zzandori.zzanmoa.item.service;
 
 import java.util.List;
 import java.util.Map;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import zzandori.zzanmoa.item.entity.Item;
+import zzandori.zzanmoa.item.repository.ItemRepository;
 import zzandori.zzanmoa.market.entity.Market;
 import zzandori.zzanmoa.market.repository.MarketRepository;
-import zzandori.zzanmoa.item.repository.ItemRepository;
 
 @RequiredArgsConstructor
 @Service
