+++ src/main/java/zzandori/zzanmoa/marketplace/controller/MarketPlaceController.java
@@ -7,14 +7,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import zzandori.zzanmoa.market.dto.MarketResponseDto;
 import zzandori.zzanmoa.marketplace.dto.MarketPlaceResponseDto;
 import zzandori.zzanmoa.marketplace.service.MarketPlaceService;
 
-@Tag(name = "MarketPlaceController", description = "시장 정보를 가져오는 컨트롤러")
+@Tag(name = "MarketPlaceController", description = "시장의 가게 정보 응답 컨트롤러")
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/market-place")
+@RequestMapping("/market/market-place")
 public class MarketPlaceController {
 
     private final MarketPlaceService marketPlaceService;
